+++ backend/infra/web/src/main/kotlin/dev/marcal/chatvault/web/ChatController.kt
@@ -1,7 +1,6 @@
 package dev.marcal.chatvault.web
 
 import dev.marcal.chatvault.in_out_boundary.input.AttachmentCriteriaInput
-import dev.marcal.chatvault.in_out_boundary.input.FileTypeInputEnum
 import dev.marcal.chatvault.in_out_boundary.output.AttachmentInfoOutput
 import dev.marcal.chatvault.in_out_boundary.output.ChatLastMessageOutput
 import dev.marcal.chatvault.in_out_boundary.output.MessageOutput
@@ -14,9 +13,6 @@ import org.springframework.data.web.SortDefault
 import org.springframework.http.*
 import org.springframework.web.bind.annotation.*
 import org.springframework.web.multipart.MultipartFile
-import org.springframework.web.server.ResponseStatusException
-import org.springframework.web.util.HtmlUtils
-import java.util.*
 import java.util.concurrent.TimeUnit
 
 @RestController
@@ -24,10 +20,7 @@ import java.util.concurrent.TimeUnit
 class ChatController(
     private val chatLister: ChatLister,
     private val messageFinderByChatId: MessageFinderByChatId,
-    private val chatFileImporter: ChatFileImporter,
     private val attachmentFinder: AttachmentFinder,
-    private val bucketDiskImporter: BucketDiskImporter,
-    private val chatFileExporter: ChatFileExporter,
     private val chatNameUpdater: ChatNameUpdater,
     private val attachmentInfoFinderByChatId: AttachmentInfoFinderByChatId,
     private val profileImageManager: ProfileImageManager
@@ -52,14 +45,6 @@ class ChatController(
         )
     }
 
-    @PostMapping("{chatId}/messages/import")
-    fun importFileByChatId(
-        @PathVariable chatId: Long,
-        @RequestParam("file") file: MultipartFile
-    ): ResponseEntity<String> {
-        return importChat(chatId = chatId, file = file)
-    }
-
     @PatchMapping("{chatId}/chatName/{chatName}")
     fun update(
         @PathVariable chatId: Long,
@@ -68,78 +53,6 @@ class ChatController(
         chatNameUpdater.execute(chatId, chatName)
         return ResponseEntity.noContent().build()
     }
-
-    @PostMapping("import/{chatName}")
-    fun importChatByChatName(
-        @PathVariable chatName: String,
-        @RequestParam("file") file: MultipartFile
-    ): ResponseEntity<String> {
-        return importChat(chatName = chatName, file = file)
-    }
-
-    @GetMapping("{chatId}/export")
-    fun importChatByChatName(
-        @PathVariable chatId: Long,
-    ): ResponseEntity<Resource> {
-        val resource = chatFileExporter.execute(chatId)
-        val headers = HttpHeaders()
-        headers.add(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=${UUID.randomUUID()}.zip")
-        headers.contentType = MediaType.APPLICATION_OCTET_STREAM
-        return ResponseEntity.ok()
-            .headers(headers)
-            .body(resource)
-    }
-
-
-    fun importChat(chatId: Long? = null, chatName: String? = null, file: MultipartFile): ResponseEntity<String> {
-        if (file.isEmpty) {
-            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "This file cannot be imported. It is empty.")
-        }
-
-        val fileType = when (file.contentType) {
-            null -> throw ResponseStatusException(
-                HttpStatus.BAD_REQUEST,
-                "This file cannot be imported. Media type is required."
-            )
-
-            "text/plain" -> FileTypeInputEnum.TEXT
-            "application/zip" -> FileTypeInputEnum.ZIP
-            else -> throw ResponseStatusException(
-                HttpStatus.UNSUPPORTED_MEDIA_TYPE,
-                "This file cannot be imported. Media type not supported ${HtmlUtils.htmlEscape(file.contentType!!)}."
-            )
-        }
-
-        importChat(chatId, file, fileType, chatName)
-        return ResponseEntity.noContent().build()
-    }
-
-    private fun importChat(
-        chatId: Long?,
-        file: MultipartFile,
-        fileType: FileTypeInputEnum,
-        chatName: String?
-    ) {
-        chatId?.let {
-            chatFileImporter.execute(
-                chatId = it,
-                inputStream = file.inputStream,
-                fileType = fileType
-            )
-        } ?: run {
-            chatFileImporter.execute(
-                chatName = chatName,
-                inputStream = file.inputStream,
-                fileType = fileType
-            )
-        }
-    }
-
-    @PostMapping("disk-import")
-    fun executeDiskImport() {
-        bucketDiskImporter.execute()
-    }
-
 
     @GetMapping("{chatId}/messages/{messageId}/attachment")
     fun downloadAttachment(
